+++ src/whatsapp/exporter.py
@@ -7,12 +7,12 @@ import json
 
 from attendanceConfig import RuntimeConfig, writeCsv
 from organiseMyProjects.logUtils import getLogger  # type: ignore[import]
-from whatsappSelectors import DEFAULT_SELECTORS, WhatsAppSelectors
+from whatsapp.selectors import DEFAULT_SELECTORS, WhatsAppSelectors
 
-from whatsappBrowser import WhatsAppPollScraper
-from whatsappCache import PollCacheStore
-from whatsappParsing import PollTextParser
-from whatsappReports import AttendanceReportBuilder
+from whatsapp.browser import WhatsAppPollScraper
+from whatsapp.cache import PollCacheStore
+from whatsapp.parsing import PollTextParser
+from whatsapp.reports import AttendanceReportBuilder
 
 logger = getLogger()
