+++ src/whatsappAttendance.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 # Compatibility wrapper. Existing imports can keep using:
 # from whatsappAttendance import AttendanceExporter, PollRecord, PollSession
 
-from whatsappExporter import AttendanceExporter
-from whatsappModels import PollRecord, PollSession
+from whatsapp.exporter import AttendanceExporter
+from whatsapp.models import PollRecord, PollSession
 
 __all__ = ["AttendanceExporter", "PollRecord", "PollSession"]
